Descriptor-based fields, linters and docs
Major project refactoring change, in which important functionality
is modified, files are reformatted and documentation is completely
rewritten.

Descriptor protocol is now used for defining fields tracked by the
framework. The built-in support for dataclasses is dropped in favor of
this.

The code now follows formatting rules imposed by Black, isort, flake8
and Darglint. The documentation is fresh and includes a lot of examples.

The version is now bumped to the first major beta.