fix: Model update and discarded models
Fixes the issue in which a model update would affect the dictionary of
persisted values while the model was not yet bound to a Transaction.
Now, the `_persistent_values` update is triggered by the Transaction
itself upon event dispatching started by the model instance.

Also, an issue with `Transaction.reset()` not marking models as
DISCARDED was fixed. A new restriction was added, in which it is not
possible anymore to register a model to any Transaction if it has been
discarded.